create receive ipfield base
